map and table appear together
